refactored the build_dataset script into race and season pipelines
