+++ pipelines/race_pipeline/extract/__init__.py
@@ -1,0 +1,5 @@
+from pipelines.race_pipeline.extract.extractor import RaceExtractor
+
+__all__ = [
+    "RaceExtractor"
+]
