+++ pipelines/race_pipeline/load/__init__.py
@@ -1,0 +1,5 @@
+from pipelines.race_pipeline.load.race_loader import RaceLoader
+
+__all__ = [
+    "RaceLoader"
+]
