+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+testpaths = pipelines/test
+pythonpath = .
